Extract the expression projection logic into a separate function
